feat: Show notes on ride
