+++ src/components/RideDetails/RideNotesForm.tsx
@@ -1,13 +1,9 @@
-import { FormEventHandler, Fragment, useState } from "react";
+import { FormEventHandler } from "react";
 import {
   UseFormRegister,
   FieldErrorsImpl,
-  useForm,
-  SubmitHandler,
+  UseFormSetValue,
 } from "react-hook-form";
-import { Dialog, Transition } from "@headlessui/react";
-import { useSWRConfig } from "swr";
-import { updateRideNotes } from "../../hooks";
 import { Button } from "../Button";
 
 export type FormValues = {
@@ -19,139 +15,55 @@ type RideNotesFormProps = {
   register: UseFormRegister<FormValues>;
   errors: Partial<FieldErrorsImpl<FormValues>>;
   handleSubmit: FormEventHandler<HTMLFormElement>;
+  setValue: UseFormSetValue<FormValues>;
   isDirty: boolean;
   waiting: boolean;
   closeHandler?: () => void;
 };
 
 // Define the form
-const RideNotesForm = ({
+export const RideNotesForm = ({
   defaultValues,
   register,
   handleSubmit,
+  setValue,
   waiting,
   isDirty,
   closeHandler,
-}: RideNotesFormProps) => (
-  <form
-    className="relative grid w-full grid-cols-1 gap-4"
-    onSubmit={handleSubmit}
-  >
-    <div className="flex flex-col gap-4 md:gap-8">
-      <label htmlFor="notes" className="flex flex-col">
-        <textarea
-          id="notes"
-          aria-label="ride note"
-          className="rounded"
-          rows={4}
-          defaultValue={defaultValues.notes}
-          {...register("notes")}
-        />
-      </label>
-    </div>
-
-    <div className="grid w-full grid-cols-2 grid-rows-[48px] gap-4 md:gap-8">
-      <Button loading={waiting} disabled={!isDirty} type="submit">
-        <div>Set Note</div>
-      </Button>
-      <Button variant="cancel" disabled={waiting} onClick={closeHandler}>
-        <span>Cancel</span>
-      </Button>
-    </div>
-  </form>
-);
-
-type Props = {
-  rideId?: string;
-  userId?: string;
-  rideNotes?: string;
-  showNotesForm: boolean;
-  closeHandler: () => void;
-};
-
-export const RideNotes = ({
-  rideId,
-  userId,
-  rideNotes,
-  showNotesForm,
-  closeHandler,
-}: Props) => {
-  const { mutate } = useSWRConfig();
-  const [waiting, setWaiting] = useState(false);
-
-  const {
-    register,
-    handleSubmit,
-    formState: { errors, isDirty },
-  } = useForm<FormValues>();
-
-  // Initial state for form: set name, leader and time
-  const defaultValues = { notes: rideNotes };
-
-  const onSubmit: SubmitHandler<FormValues> = async ({ notes }) => {
-    setWaiting(true);
-    if (rideId && userId) {
-      await mutate(`/api/ride/${rideId}`, async () => {
-        // Await update and subsequent requery, to refresh notes in page
-        await updateRideNotes(rideId, userId, notes || "");
-      });
-    }
-    setWaiting(false);
-    closeHandler();
+}: RideNotesFormProps) => {
+  const clearForm = () => {
+    setValue("notes", "");
   };
 
   return (
-    <Transition appear show={showNotesForm} as={Fragment}>
-      <Dialog as="div" className="relative z-10" onClose={closeHandler}>
-        <Transition.Child
-          as={Fragment}
-          enter="ease-out duration-300"
-          enterFrom="opacity-0"
-          enterTo="opacity-100"
-          leave="ease-in duration-200"
-          leaveFrom="opacity-100"
-          leaveTo="opacity-0"
-        >
-          <div className="fixed inset-0 bg-black bg-opacity-25" />
-        </Transition.Child>
+    <form
+      className="relative grid w-full grid-cols-1 gap-4"
+      onSubmit={handleSubmit}
+    >
+      <div className="flex flex-col gap-4 md:gap-8">
+        <label htmlFor="notes" className="flex flex-col">
+          <textarea
+            id="notes"
+            aria-label="ride note"
+            className="rounded"
+            rows={4}
+            defaultValue={defaultValues.notes}
+            {...register("notes")}
+          />
+        </label>
+      </div>
 
-        <div className="fixed inset-0 overflow-y-auto">
-          <div className="flex min-h-full items-center justify-center p-4 text-center">
-            <Transition.Child
-              as={Fragment}
-              enter="ease-out duration-300"
-              enterFrom="opacity-0 scale-95"
-              enterTo="opacity-100 scale-100"
-              leave="ease-in duration-200"
-              leaveFrom="opacity-100 scale-100"
-              leaveTo="opacity-0 scale-95"
-            >
-              <Dialog.Panel className="w-full max-w-md transform overflow-hidden rounded bg-white p-6 text-left align-middle shadow-xl transition-all">
-                <Dialog.Title
-                  as="h3"
-                  className="text-lg font-medium leading-6 text-gray-900"
-                >
-                  Add ride notes
-                </Dialog.Title>
-
-                <div className="mt-2">
-                  <div className="w-full text-neutral-800">
-                    <RideNotesForm
-                      defaultValues={defaultValues}
-                      errors={errors}
-                      isDirty={isDirty}
-                      register={register}
-                      handleSubmit={handleSubmit(onSubmit)}
-                      waiting={waiting}
-                      closeHandler={closeHandler}
-                    />
-                  </div>
-                </div>
-              </Dialog.Panel>
-            </Transition.Child>
-          </div>
-        </div>
-      </Dialog>
-    </Transition>
+      <div className="grid w-full grid-cols-3 grid-rows-[48px] gap-4 md:gap-8">
+        <Button loading={waiting} disabled={!isDirty} type="submit">
+          <div>Set</div>
+        </Button>
+        <Button onClick={clearForm} type="submit">
+          <div>Clear</div>
+        </Button>
+        <Button variant="cancel" disabled={waiting} onClick={closeHandler}>
+          <span>Cancel</span>
+        </Button>
+      </div>
+    </form>
   );
 };
